+++ finance_digest/lib/components/news_card.dart
@@ -1,5 +1,5 @@
 import 'package:finance_digest/modules/models/news.dart';
-import 'package:finance_digest/widgets/spacers.dart';
+import 'package:finance_digest/widgets/helper_functions.dart';
 import 'package:flutter/material.dart';
 
 class NewsCard extends StatelessWidget {
