+++ finance_digest/lib/components/no_news.dart
@@ -1,4 +1,4 @@
-import 'package:finance_digest/widgets/spacers.dart';
+import 'package:finance_digest/widgets/helper_functions.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
