+++ finance_digest/lib/modules/models/web_view.dart
@@ -1,0 +1,7 @@
+// import 'package:webview_flutter/webview_flutter.dart';
+
+// class WebView {
+//   final String title;
+//   final WebViewController controller;
+//   WebView({required this.title, required this.controller});
+// }
